requirements.txt
openai>=1.30.1
python-dotenv>=1.0.1
rich>=13.7.0
